solved problem with subtask click event listener
When a task is generated, a "click" eventlistener is added, which executes a function editSubtask() when the task gets edited. Once the task gets edited, the "click" eventlistener for calling the editSubtask() gets removed and a new "click" eventlistener is added to the subtask, which call the saveSubtaskChanged() function. In this function, the last eventlistener for saveSubtaskChanges() will be removed and the editSubtask() eventlistener gets added like default.