fixed bug on drag and drop event
It's not possible anymore, to drop tasks in other task. Solved over a class. Each task has class noDrop on every element added in a task. If you drag and drop a task on another, ev.target will check, if the target has the class noDrop. If yes, ev.preventDefault() will be called.