Made a few css changes
Break the line in the div which shows the added users to a tasks. Used flex-wrap: wrap; so the content can't go beyon a task.

Break the word in a contact in the e-mail adress field, so the content can't go beyond a contact.

Added margin-left to the urgent date in the summary so the date cant go over the urgend tasks field.